refactor(filters): one copy of the date filter operators
`DataTable` and `DataView` each carried their own `filter-operations.tsx` with
the same six date operators and the same `compare` helper — 34 of the 40 lines
this PR added to each were byte-identical, comment included, and the
epoch-seconds fix had to be made twice by hand.

Nothing about the split was load-bearing: both already type the map as
tanstack's `FilterFn<unknown>` and read `date` off the same `~/types/filters`
`FilterValue`. The copies differed only in arrow-body style and an unused
`_addMeta`.

Sited beside `date-adapter` rather than under `shared/`. Both modules already
import the adapter from there, so this adds no new dependency edge, whereas
`shared/` is a leaf every component depends on and must not depend back on one
— the first draft put it there and inverted that. Kept out of `date-adapter`
itself so a tanstack `FilterFn` never joins the calendar's date surface.

  identical added lines  34 → 5
  divergence            181 → 147 lines
  combined size         551 → 455 lines

The remaining divergence is the string, number and select operators, which
genuinely differ between the two packages.
